requirements.txt
Flask==2.1.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-Security-Too==4.1.1
Flask-Admin==1.6.0
Flask-Mail==0.9.1
passlib==1.7.4
schedule==1.1.0
Flask-JWT-Extended==4.3.1
Flask-Migrate==3.1.0
pymysql==1.0.2
PyMySQL[rsa]==1.0.2
email-validator==1.1.3
ansible==4.8.0
